Do project restructure, merge rank extension and service
