+++ src/db/converters.py
@@ -13,6 +13,7 @@
 
 __author__ = 'Mathtin'
 
+from typing import Any, Dict, List
 import discord as d
 from .models import User, MessageEvent
 from datetime import datetime
@@ -21,21 +22,21 @@ from datetime import datetime
 # Roles
 #
 
-def role_to_row(role: d.Role):
+def role_to_row(role: d.Role) -> Dict[str, Any]:
     return {
         'did': role.id,
         'name': role.name,
         'created_at': role.created_at
     }
 
-def roles_to_rows(roles: list):
+def roles_to_rows(roles: list) -> List[Dict[str, Any]]:
     rows = [role_to_row(r) for r in roles]
     rows = sorted(rows, key = lambda i: i['did'])
     for i in range(len(rows)):
         rows[i]['idx'] = i
     return rows
 
-def role_mask(user: d.Member, role_map: dict):
+def role_mask(user: d.Member, role_map: dict) -> str:
     mask = ['0'] * len(role_map)
     for role in user.roles:
         idx = role_map[role.id]['idx']
@@ -46,7 +47,7 @@ def role_mask(user: d.Member, role_map: dict):
 # Users
 #
 
-def user_row(user: d.User):
+def user_row(user: d.User) -> Dict[str, Any]:
     return {
         'did': user.id,
         'name': user.name,
@@ -55,7 +56,7 @@ def user_row(user: d.User):
         'roles': None
     }
 
-def member_row(user: d.Member, role_map: dict):
+def member_row(user: d.Member, role_map: dict) -> Dict[str, Any]:
     return {
         'did': user.id,
         'name': user.name,
@@ -65,14 +66,14 @@ def member_row(user: d.Member, role_map: dict):
         'created_at': user.joined_at
     }
 
-def member_join_row(user: User, joined: datetime ,events: dict):
+def member_join_row(user: User, joined: datetime ,events: dict) -> Dict[str, Any]:
     return {
         'type_id': events["member_join"],
         'user_id': user.id,
         'created_at': joined
     }
 
-def user_leave_row(user: User, events: dict):
+def user_leave_row(user: User, events: dict) -> Dict[str, Any]:
     return {
         'type_id': events["member_leave"],
         'user_id': user.id
@@ -81,7 +82,7 @@ def user_leave_row(user: User, events: dict):
 # Messages
 #
 
-def new_message_to_row(user_id: int, msg: d.Message, events: dict):
+def new_message_to_row(user_id: int, msg: d.Message, events: dict) -> Dict[str, Any]:
     return {
         'type_id': events["new_message"],
         'user_id': user_id,
@@ -90,7 +91,7 @@ def new_message_to_row(user_id: int, msg: d.Message, events: dict):
         'created_at': msg.created_at
     }
 
-def message_edit_row(msg: MessageEvent, events: dict):
+def message_edit_row(msg: MessageEvent, events: dict) -> Dict[str, Any]:
     return {
         'type_id': events["message_edit"],
         'user_id': msg.user.id,
@@ -98,7 +99,7 @@ def message_edit_row(msg: MessageEvent, events: dict):
         'channel_id': msg.channel_id
     }
 
-def message_delete_row(msg: MessageEvent, events: dict):
+def message_delete_row(msg: MessageEvent, events: dict) -> Dict[str, Any]:
     return {
         'type_id': events["message_delete"],
         'user_id': msg.user.id,
@@ -110,14 +111,14 @@ def message_delete_row(msg: MessageEvent, events: dict):
 # VC
 #
 
-def vc_join_row(user: User, channel: d.VoiceChannel, events: dict):
+def vc_join_row(user: User, channel: d.VoiceChannel, events: dict) -> Dict[str, Any]:
     return {
         'type_id': events["vc_join"],
         'user_id': user.id,
         'channel_id': channel.id
     }
 
-def vc_leave_row(user: User, channel: d.VoiceChannel, events: dict):
+def vc_leave_row(user: User, channel: d.VoiceChannel, events: dict) -> Dict[str, Any]:
     return {
         'type_id': events["vc_leave"],
         'user_id': user.id,
@@ -128,7 +129,7 @@ def vc_leave_row(user: User, channel: d.VoiceChannel, events: dict):
 # User Stat
 #
 
-def empty_user_stat_row(user_id: int, type_id: int):
+def empty_user_stat_row(user_id: int, type_id: int) -> Dict[str, Any]:
     return {
         'type_id': type_id,
         'user_id': user_id,
